backend/pyproject.toml
[project]
name = "avatar-adam-be"
version = "0.1.0"
description = "FastAPI project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    # Database & ORM
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    # Configuration
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    # Utilities
    "tenacity>=8.2.3",
    # AI/ML - LangChain & RAG
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    # Vector Database - pgvector (PostgreSQL extension)
    "pgvector>=0.3.0",
    # Vector Database - Pinecone
    "pinecone>=5.0.0",
    # LLM - OpenRouter (uses OpenAI SDK) and STT (Whisper)
    "openai>=1.0.0",
    # Document processing
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    # Async utilities
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
    # WebSocket client for ElevenLabs streaming TTS
    "websockets>=12.0",
    # Email - Mailgun SDK (pip install mailgun)
    "mailgun>=1.5.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Password validation for bcrypt
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "faker>=20.1.0",
]
